backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.2.0
torch>=2.0.0
